Rename routes and components so that the router has only route files
