--- src/components/App.jsx
@@ -1,15 +1,0 @@
-import '@/styles/App.css'
-import Header from '@/components/Header'
-import ForecastDetails from '@/components/ForecastDetails'
-import MountainsList from '@/components/MountainsList'
-
-function App() {
-  return (
-    <>
-      <Header />
-      <MountainsList />
-    </>
-  )
-}
-
-export default App
