+++ src/components/MountainsList.jsx
@@ -1,11 +1,8 @@
-import { useSelector, useDispatch } from 'react-redux';
 import styles from '@/styles/MountainsList.module.css';
 import { Link } from "react-router-dom";
 
-const MountainsList = () => {
-  const dispatch = useDispatch();
-  const { mountains, isLoading, error } = useSelector((store) => store.mountains);
-
+const MountainsList = ({data}) => {
+  const { mountains, isLoading, error } = data;
   if (isLoading) {
     return <p className={styles.loading}>The forecast is loading!</p>;
   }
@@ -22,13 +19,13 @@ const MountainsList = () => {
     return (
       <div>
         {
-          mountains.map((peak) => (
-            <>
+          data.mountains.map((peak) => (
+            <li key={peak.name}>
               <p>{peak.name}</p>
               <p>{peak.elevation}</p>
               <p>{JSON.stringify(peak.current_weather)}</p>
               <Link to={`details/${peak.name}`}>See Details -&gt;</Link>
-            </>
+            </li>
           ))
         }
       </div>
