+++ src/main.jsx
@@ -1,14 +1,14 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import Home from './routes/Home';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import Root from './routes/Root';
 import ErrorPage from "./components/ErrorPage";
 import Details from './routes/Details';
 import {
   createBrowserRouter,
   RouterProvider,
 } from "react-router-dom";
-import './index.css'
-import MountainsList from './components/MountainsList';
+import './index.css';
+import Home from './routes/Home';
 import About from './components/About';
 import store from '@/redux/store';
 import { Provider } from 'react-redux';
@@ -16,7 +16,7 @@ import { Provider } from 'react-redux';
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <Home />,
+    element: <Root />,
     errorElement: <ErrorPage />,
     children: [
       {
@@ -25,7 +25,7 @@ const router = createBrowserRouter([
       },
       {
         path: "/",
-        element: <MountainsList />,
+        element: <Home />,
       },
       {
         path: "about",
